backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
anthropic==0.34.2
google-auth==2.35.0
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
apscheduler==3.10.4
httpx==0.27.2
python-dotenv==1.0.1
